+++ src/components/domain/CommentItem/index.ts
@@ -1,0 +1,1 @@
+export { default as CommentItem } from './CommentItem';
