+++ src/components/domain/index.ts
@@ -2,3 +2,5 @@ export * from './PostCard';
 export * from './List';
 export * from './Footer';
 export * from './Header';
+export * from './PostList';
+export * from './CommentItem';
